+++ index.html
@@ -2,7 +2,10 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
+    <meta name="description" content="Canned responses" />
+    <meta name="keywords" content="canned, responses, random, user" />
     <link rel="icon" type="image/svg+xml" href="/vite.svg" />
+    <link rel="preconnect" href="https://randomuser.me">
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Canned responses</title>
   </head>
