+++ src/components/canned-responses/author/styles.ts
@@ -1,0 +1,23 @@
+import { css } from '@emotion/css';
+
+export const author = css`
+  display: flex;
+`;
+
+export const avatarContainer = css`
+  margin-right: 8px;
+`;
+
+export const authorNameSkeleton = css`
+  margin: 4px;
+`;
+
+export const authorName = css`
+  margin-left: 4px;
+`;
+
+export const avatar = css`
+  display: inline-block;
+  vertical-align: middle;
+  user-select: text;
+`;
